graphs/requirements.txt
Django==1.10
Jinja2==2.7.2
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.2.1
boto==2.20.1
chardet==2.0.1
click==6.2
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
decorator==4.0.6
defer==1.0.6
dirspec==13.10
djangorestframework==3.4.3
dnspython==1.11.1
duplicity==0.6.23
et-xmlfile==1.0.1
geocoder==1.6.4
gyp==0.1
html5lib==0.999
httplib2==0.8
ipython==1.2.1
jdcal==1.2
lockfile==0.8
lxml==3.3.3
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
openpyxl==2.3.2
pexpect==3.1
piston-mini-client==0.7.5
plotly==1.8.12
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu2
python-cloudfiles==1.7.10
python-debian==0.1.21-nmu2ubuntu2
pytz==2015.7
pyxdg==0.25
pyzmq==14.0.1
ratelim==0.1.6
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplekml==1.2.8
six==1.5.2
software-center-aptd-plugins==0.0.0
system-service==0.1.6
tornado==3.1.1
urllib3==1.7.1
virtualenv==15.0.3
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
xlrd==0.9.4
zope.interface==4.0.5
